requirements.txt
Flask==2.3.3
pymongo==4.5.0
Werkzeug==2.3.7
cloudinary==1.34.0
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
